+++ src/libs6/s6_accessrules_keycheck_reversedns.c
@@ -1,5 +1,6 @@
 /* ISC license. */
 
+#include <sys/types.h>
 #include <errno.h>
 #include <skalibs/bytestr.h>
 #include <s6/accessrules.h>
@@ -7,11 +8,11 @@
 s6_accessrules_result_t s6_accessrules_keycheck_reversedns (void const *key, void *data, s6_accessrules_params_t *params, s6_accessrules_backend_func_t_ref check1)
 {
   char const *name = key ;
-  unsigned int len = str_len(name) ;
+  size_t len = str_len(name) ;
   if (!len) return (errno = EINVAL, S6_ACCESSRULES_ERROR) ;
   if (name[len-1] == '.') len-- ;
   {
-    unsigned int i = 0 ;
+    size_t i = 0 ;
     char tmp[len + 11] ;
     byte_copy(tmp, 11, "reversedns/") ;
     while (i < len)
